requirements.txt
# --- Core Web / Auth / DB ---
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Cors==6.0.1
python-dotenv==1.2.1
SQLAlchemy==2.0.44
alembic==1.17.1
psycopg2-binary==2.9.11
greenlet==3.2.4

# --- Scientific / RF stack ---
numpy==2.1.3
scipy==1.14.1
matplotlib==3.10.7
Pillow==12.0.0
astropy==7.1.1
pycraf==2.1.0
pyproj==3.7.2
rasterio==1.4.3
shapely==2.1.2
geopy==2.4.1
geojson==3.2.0
scikit-learn==1.7.2
reportlab==4.4.4

# --- Utilities & helpers ---
requests==2.32.5
python-dateutil==2.9.0.post0
rich==14.2.0
PyYAML==6.0.3
passlib[argon2]==1.7.4
argon2-cffi==25.1.0
joblib==1.5.2
threadpoolctl==3.6.0
pdfplumber==0.11.8
google-generativeai==0.8.5
